LayoutTests/media/modern-media-controls/ios-inline-media-controls/ios-inline-media-controls-layout-expected.txt

Testing the IOSInlineMediaControls layout.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.controlsBar.children[0] instanceof ButtonsContainer is true
PASS mediaControls.controlsBar.children[1] instanceof LayoutNode is true
PASS mediaControls.controlsBar.children[2] instanceof ButtonsContainer is true
PASS mediaControls.timeControl.x is 118
PASS mediaControls.timeControl.width is 393
PASS mediaControls.controlsBar.children[0].children is [mediaControls.playPauseButton, mediaControls.skipBackButton]
PASS mediaControls.controlsBar.children[2].children is [mediaControls.airplayButton, mediaControls.pipButton, mediaControls.fullscreenButton]
PASS successfullyParsed is true

TEST COMPLETE
